Corrected Comment style, changed method name added Append switch
Changed the comments to correct format for Visual Studio methods.
Changed the name of the AddEvent method to AddLog to make it clear in
case using the EventManager at the same time as the LogManager.
Added ability to force append, although it still respects large files
and handles in the same method

+++ LogFileManager.cs
@@ -11,11 +11,14 @@ namespace LFManager
         private StreamWriter sw = null;
         private string logFileLoc = null;
 
-        /*
-         * Create the Event Source (if required). 
-         * Create a header to the event log file for each time the service is started.
-         */ 
-        public LogFileManager(string app, string logPath = "")
+        /// <summary>
+        /// Create the Event Source (if required). 
+        /// Create a header to the event log file for each time the service is started.
+        /// </summary>
+        /// <param name="app">Application Name, forms part of the log file name</param>
+        /// <param name="logPath">If specfied will attempt to create the log file at this location otherwise uses currently executable's location</param>
+        /// <param name="append">Force if the file is appeneded to or not</param>
+        public LogFileManager(string app, string logPath = "", bool append = false)
         {
             // If the path to the log file has not been passed in use the location of the exe
             if (logPath == "")
@@ -25,37 +28,44 @@ namespace LFManager
                 logPath += "\\" + app + "_Log.txt";                 // adds the required \ and our log file
             }
 
-            try // Access an existing file
+            if (append) // Allow user to force append for logging file, but adhere to archiving large files
             {
-                FileInfo file = new FileInfo(logPath);
-    
-                // Check to see if the log file is getting too large (1 Megabyte)
-                if (file.Length > 1000000)
+                try // Access an existing file
                 {
-                    // Move existing (large file) to an archived file and start a new log
-                    
-                    // Build new file name using the date
-                    DateTime date = new DateTime();
-                    date = DateTime.Today;
-                    
-                    string archivePath = Assembly.GetExecutingAssembly().Location;  // See comments on line 23!
-                    archivePath = Path.GetDirectoryName(logPath);
-                    archivePath += app + "_Archived_" + date.ToString() + ".txt";
-    
-                    file.CopyTo(archivePath);
+                    FileInfo file = new FileInfo(logPath);
+
+                    // Check to see if the log file is getting too large (1 Megabyte)
+                    if (file.Length > 1000000)
+                    {
+                        // Move existing (large file) to an archived file and start a new log
+
+                        // Build new file name using the date
+                        DateTime date = new DateTime();
+                        date = DateTime.Today;
+
+                        string archivePath = Assembly.GetExecutingAssembly().Location;  // See comments on line 23!
+                        archivePath = Path.GetDirectoryName(logPath);
+                        archivePath += app + "_Archived_" + date.ToString() + ".txt";
+
+                        file.CopyTo(archivePath);
+                        sw = new StreamWriter(logPath, false);
+                    }
+                    else // continue adding to file
+                    {
+                        sw = new StreamWriter(logPath, true);
+                    }
+                }
+                catch (IOException)
+                {
+                    // File didn't exist, set up a new log file StreamWriter
                     sw = new StreamWriter(logPath, false);
                 }
-                else // continue adding to file
-                {
-                    sw = new StreamWriter(logPath, true);
-                }
             }
-            catch (IOExecption)
+            else // Do not append and always start a new file
             {
-                // File didn't exist, set up a new log file StreamWriter
                 sw = new StreamWriter(logPath, false);
             }
-            
+
             try
             {
                 // Write to the log stipulating the application was started
@@ -73,15 +83,16 @@ namespace LFManager
             logFileLoc = logPath;
         }
 
-        /*
-         * Provide a method for adding a new event to the log file
-         */ 
-        public void AddEvent(string logEvent)
+        /// <summary>
+        /// Provides a quick way to write to the file, always appends to the created log file in the constructor
+        /// </summary>
+        /// <param name="logMsg">The message/details to enter in to the log file</param>
+        public void AddLog(string logMsg)
         {
             try // write to the log file
             {
                 sw = new StreamWriter(logFileLoc, true);
-                sw.WriteLine(logEvent);
+                sw.WriteLine(logMsg);
                 sw.Close();
             }
             catch
